chore: added namespace adjustments for messaging classes
