+++ libs/sdtechbank/SdtechBank.Application/IntegrationEvents/IIntegrationEventDispatcher.cs
@@ -1,0 +1,8 @@
+namespace SdtechBank.Application.IntegrationEvents;
+
+public interface IIntegrationEventDispatcher
+{
+    Task DispatchAsync<TEvent>(TEvent @event, CancellationToken cancellationToken);
+}
+
+
