--- libs/sdtechbank/SdtechBank.Application/IntegrationEvents/IOutboxService.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-
-namespace SdtechBank.Application.Messaging;
-
-public interface IOutboxService
-{
-    Task AddEventAsync<T>(T @event, CancellationToken ct);
-}
